# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACES MSD Revenue Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: List revenue records with filters
      parameters:
        - name: project
          in: query
          schema: { type: string }
        - name: projectStatus
          in: query
          schema: { type: string }
        - name: revenueYear
          in: query
          schema: { type: ["number", "null"] }
        - name: revenueMonth
          in: query
          schema: { type: ["number", "null"] }
        - name: invoiceStatus
          in: query
          schema: { type: string }
        - name: invoiceNo
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
        - name: overdue
          in: query
          schema: { type: ["boolean", "null"] }
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: pageSize
          in: query
          schema: { type: number, default: 50 }
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordListResponse"
    post:
      operationId: createRecord
      tags: [records]
      summary: Create a revenue record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueRecord"
  /records/import:
    post:
      operationId: importRecords
      tags: [records]
      summary: Import records from Excel/CSV (JSON parsed client-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPayload"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a revenue record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueRecord"
        "404":
          description: Not found
    patch:
      operationId: updateRecord
      tags: [records]
      summary: Update a revenue record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteRecord
      tags: [records]
      summary: Delete a revenue record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    RevenueRecord:
      type: object
      required: [id, projectName, revenueMonth, workOrder, revenue, deductible, invoiced, collected, penalties, netRevenue, paymentStatus, isDemo]
      properties:
        id: { type: number }
        projectId: { type: ["number", "null"] }
        projectName: { type: string }
        revenueMonth: { type: string, description: "ISO date YYYY-MM-DD (first day of month)" }
        workOrder: { type: number }
        revenue: { type: number }
        deductible: { type: number }
        invoiced: { type: number }
        invoiceDate: { type: ["string", "null"] }
        invoiceNo: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        collected: { type: number }
        collectedDate: { type: ["string", "null"] }
        days: { type: ["number", "null"] }
        penalties: { type: number }
        netRevenue: { type: number }
        paymentStatus: { type: string }
        outstandingAmount: { type: number }
        overdueAmount: { type: number }
        outstandingAgeDays: { type: ["number", "null"] }
        deductibleVariance: { type: ["number", "null"] }
        netRevenueVariance: { type: ["number", "null"] }
        retentionApplicable: { type: boolean }
        releasePercentage: { type: number }
        initialReleaseAmount: { type: number }
        retainedAmount: { type: number }
        retentionStatus: { type: ["string", "null"], description: "Withheld | Eligible for Release | Invoiced | Collected" }
        pendingRetention: { type: number }
        bodStatus: { type: ["string", "null"], description: "Pending | Submitted | Approved | Signed" }
        bodCompletionDate: { type: ["string", "null"] }
        retentionReleaseDate: { type: ["string", "null"] }
        isDemo: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    RecordInput:
      type: object
      required: [projectName, revenueMonth, workOrder, revenue]
      properties:
        projectName: { type: string }
        revenueMonth: { type: string }
        workOrder: { type: number }
        revenue: { type: number }
        deductible: { type: number }
        invoiced: { type: number }
        invoiceDate: { type: ["string", "null"] }
        invoiceNo: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        collected: { type: number }
        collectedDate: { type: ["string", "null"] }
        days: { type: ["number", "null"] }
        penalties: { type: number }
        netRevenue: { type: number }
        bodStatus: { type: ["string", "null"] }
        bodCompletionDate: { type: ["string", "null"] }
        retentionReleaseDate: { type: ["string", "null"] }
        isDemo: { type: boolean }
    RecordUpdate:
      type: object
      properties:
        projectName: { type: string }
        revenueMonth: { type: string }
        workOrder: { type: number }
        revenue: { type: number }
        deductible: { type: number }
        invoiced: { type: number }
        invoiceDate: { type: ["string", "null"] }
        invoiceNo: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        collected: { type: number }
        collectedDate: { type: ["string", "null"] }
        days: { type: ["number", "null"] }
        penalties: { type: number }
        netRevenue: { type: number }
        bodStatus: { type: ["string", "null"] }
        bodCompletionDate: { type: ["string", "null"] }
        retentionReleaseDate: { type: ["string", "null"] }
    RecordListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenueRecord"
        total: { type: number }
        page: { type: number }
        pageSize: { type: number }
    ImportPayload:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/RecordInput"
        allowDuplicateInvoices: { type: boolean }
    ImportResult:
      type: object
      required: [imported, skipped, warnings, errors]
      properties:
        imported: { type: number }
        skipped: { type: number }
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ImportWarning"
        errors:
          type: array
          items: { type: string }
    ImportWarning:
      type: object
      required: [row, field, message, severity]
      properties:
        row: { type: number }
        field: { type: string }
        message: { type: string }
        severity: { type: string, enum: [warning, error] }
...