# lib/api-zod/src/generated/types/recordInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ACES MSD Revenue Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordInput {
  projectName: string;
  revenueMonth: string;
  workOrder: number;
  revenue: number;
  deductible?: number;
  invoiced?: number;
  /** @nullable */
  invoiceDate?: string | null;
  /** @nullable */
  invoiceNo?: string | null;
  /** @nullable */
  dueDate?: string | null;
  collected?: number;
  /** @nullable */
  collectedDate?: string | null;
  /** @nullable */
  days?: number | null;
  penalties?: number;
  netRevenue?: number;
  /** @nullable */
  bodStatus?: string | null;
  /** @nullable */
  bodCompletionDate?: string | null;
  /** @nullable */
  retentionReleaseDate?: string | null;
  isDemo?: boolean;
}